Move CodeMirrorDotrain component in ui-components
