+++ tauri-app/src/routes/orders/add/+page.svelte
@@ -1,7 +1,5 @@
 <script lang="ts">
-  import { PageHeader } from '@rainlanguage/ui-components';
-  import CodeMirrorDotrain from '$lib/components/CodeMirrorDotrain.svelte';
-  import { ButtonLoading } from '@rainlanguage/ui-components';
+  import { PageHeader, CodeMirrorDotrain, ButtonLoading } from '@rainlanguage/ui-components';
   import FileTextarea from '$lib/components/FileTextarea.svelte';
   import { Label, Button, Spinner, Tabs, TabItem } from 'flowbite-svelte';
   import { RawRainlangExtension, type Problem } from 'codemirror-rainlang';
@@ -40,6 +38,7 @@
   import { getAuthoringMetaV2ForScenarios } from '$lib/services/authoringMeta';
   import RaindexVersionValidator from '$lib/components/RaindexVersionValidator.svelte';
   import { page } from '$app/stores';
+  import { codeMirrorTheme } from '$lib/stores/darkMode';
 
   let isSubmitting = false;
   let isCharting = false;
@@ -220,6 +219,7 @@
 
   <svelte:fragment slot="textarea">
     <CodeMirrorDotrain
+      codeMirrorTheme={$codeMirrorTheme}
       bind:value={$globalDotrainFile.text}
       disabled={isSubmitting}
       styles={{ '&': { minHeight: '400px' } }}
